+++ app/Http/Requests/StoreOrderItems.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace KushyApi\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class StoreOrderItems extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return $this->user()->isBusiness;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'order_id'          => 'string|max:255',
+            'inventory_id'          => 'string|max:255',
+            'product_type'   => 'string',
+            'count'     => 'integer',
+        ];
+    }
+}